# Conflicts:
#	local/o365/classes/adminsetting/sdsfieldmap.php
#	local/o365/classes/feature/sds/task/sync.php
#	local/o365/classes/observers.php
#	local/o365/classes/rest/sds.php
#	local/o365/tests/sdssync_test.php

+++ local/o365/classes/feature/usergroups/utils.php
@@ -632,7 +632,9 @@ class utils {
         } else {
             if (isset($usergroupconfig[$courseid])) {
                 unset($usergroupconfig[$courseid]);
-                static::delete_course_group($courseid);
+                if (get_config('local_o365', 'delete_group_on_course_sync_disabled')) {
+                    static::delete_course_group($courseid);
+                }
             }
         }
